+++ src/pyodide-py/pyodide/console.py
@@ -1,8 +1,40 @@
-from typing import Optional, Callable
+from typing import Optional, Callable, Any
 import code
 import io
 import sys
 import platform
+from contextlib import contextmanager
+import builtins
+
+# this import can fail when we are outside a browser (e.g. for tests)
+try:
+    import js
+
+    _dummy_promise = js.Promise.resolve()
+    _load_packages_from_imports = js.pyodide.loadPackagesFromImports
+
+except ImportError:
+
+    class _FakePromise:
+        """A promise that mimic the JS promises.
+
+        Only `then is supported` and there is no asynchronicity.
+        execution occurs when then is call.
+
+        This is mainly to fake `load_packages_from_imports`
+        and `InteractiveConsole.run_complete` in contexts
+        where JS promises are not available (tests)."""
+
+        def __init__(self, args=None):
+            self.args = (args,) if args is not None else ()
+
+        def then(self, func, *args):
+            return _FakePromise(func(*self.args))
+
+    _dummy_promise = _FakePromise()
+
+    def _load_packages_from_imports(*args):
+        return _dummy_promise
 
 
 __all__ = ["InteractiveConsole"]
@@ -62,7 +94,9 @@ class InteractiveConsole(code.InteractiveConsole):
     """Interactive Pyodide console
 
     Base implementation for an interactive console that manages
-    stdout/stderr redirection.
+    stdout/stderr redirection. Since packages are loaded before running
+    code, `runcode` returns a JS promise. Override `sys.displayhook` to
+    catch the result of an execution.
 
     `self.stdout_callback` and `self.stderr_callback` can be overloaded.
 
@@ -91,12 +125,16 @@ class InteractiveConsole(code.InteractiveConsole):
         self._stderr = None
         self.stdout_callback = stdout_callback
         self.stderr_callback = stderr_callback
-        self._persistent_stream_redirection = persistent_stream_redirection
-        if self._persistent_stream_redirection:
+        self._streams_redirected = False
+        if persistent_stream_redirection:
             self.redirect_stdstreams()
+        self.run_complete = _dummy_promise
 
     def redirect_stdstreams(self):
         """ Toggle stdout/stderr redirections. """
+        # already redirected?
+        if self._streams_redirected:
+            return
 
         if self._stdout is None:
             # we use meta callbacks to allow self.std{out,err}_callback
@@ -127,27 +165,126 @@ class InteractiveConsole(code.InteractiveConsole):
         # actual redirection
         sys.stdout = self._stdout
         sys.stderr = self._stderr
+        self._streams_redirected = True
 
     def restore_stdstreams(self):
         """Restore stdout/stderr to the value it was before
         the creation of the object."""
-        sys.stdout = self._old_stdout
-        sys.stderr = self._old_stderr
+        if self._streams_redirected:
+            sys.stdout = self._old_stdout
+            sys.stderr = self._old_stderr
+            self._streams_redirected = False
 
-    def runcode(self, code):
-        if self._persistent_stream_redirection:
-            super().runcode(code)
+    @contextmanager
+    def stdstreams_redirections(self):
+        """Ensure std stream redirection.
+
+        This supports nesting."""
+        if self._streams_redirected:
+            yield
         else:
             self.redirect_stdstreams()
-            super().runcode(code)
+            yield
             self.restore_stdstreams()
 
+    def flush_all(self):
+        """ Force stdout/stderr flush. """
+        with self.stdstreams_redirections():
+            sys.stdout.flush()
+            sys.stderr.flush()
+
+    def runsource(self, *args, **kwargs):
+        """Force streams redirection.
+
+        Syntax errors are not thrown by runcode but here in runsource.
+        This is why we force redirection here since doing twice
+        is not an issue."""
+
+        with self.stdstreams_redirections():
+            return super().runsource(*args, **kwargs)
+
+    def runcode(self, code):
+        """Load imported packages then run code, async.
+
+        To achieve nice result representation, the interactive console
+        is fully implemented in Python. This has a major drawback:
+        packages should be loaded from here. This is why this
+        function sets the promise `self.run_complete`.
+        If you need to wait for the end of the computation,
+        you should await for it."""
+        parent_runcode = super().runcode
+        source = "\n".join(self.buffer)
+
+        def load_packages_and_run(*args):
+            def run(*args):
+                with self.stdstreams_redirections():
+                    parent_runcode(code)
+                    # in CPython's REPL, flush is performed
+                    # by input(prompt) at each new prompt ;
+                    # since we are not using input, we force
+                    # flushing here
+                    self.flush_all()
+                return _dummy_promise
+
+            return _load_packages_from_imports(source).then(run)
+
+        self.run_complete = self.run_complete.then(load_packages_and_run)
+
     def __del__(self):
-        if self._persistent_stream_redirection:
-            self.restore_stdstreams()
+        self.restore_stdstreams()
 
     def banner(self):
         """ A banner similar to the one printed by the real Python interpreter. """
         # copyied from https://github.com/python/cpython/blob/799f8489d418b7f9207d333eac38214931bd7dcc/Lib/code.py#L214
         cprt = 'Type "help", "copyright", "credits" or "license" for more information.'
-        return f"Python {platform.python_version()} {platform.python_build()} on WebAssembly VM\n{cprt}"
+        version = platform.python_version()
+        build = f"({', '.join(platform.python_build())})"
+        return f"Python {version} {build} on WebAssembly VM\n{cprt}"
+
+
+def repr_shorten(
+    value: Any, limit: int = 1000, split: Optional[int] = None, separator: str = "..."
+):
+    """Compute the string representation of `value` and shorten it
+    if necessary.
+
+    If it is longer than `limit` then return the firsts `split`
+    characters and the last `split` characters seperated by '...'.
+    Default value for `split` is `limit // 2`.
+    """
+    if split is None:
+        split = limit // 2
+    text = repr(value)
+    if len(text) > limit:
+        text = f"{text[:split]}{separator}{text[-split:]}"
+    return text
+
+
+def displayhook(value, repr: Callable[[Any], str]):
+    """A displayhook with custom `repr` function.
+
+    It is intendend to overload `sys.displayhook`. Note that monkeypatch
+    `builtins.repr` does not work in `sys.displayhook`. The pointer to
+    `repr` seems hardcoded in default `sys.displayhook` version
+    (which is written in C)."""
+    # from https://docs.python.org/3/library/sys.html#sys.displayhook
+    # If value is not None, this function prints repr(value) to
+    # sys.stdout, and saves value in builtins._. If repr(value) is not
+    # encodable to sys.stdout.encoding with sys.stdout.errors error
+    # handler (which is probably 'strict'), encode it to
+    # sys.stdout.encoding with 'backslashreplace' error handler.
+    if value is None:
+        return
+    builtins._ = None  # type: ignore
+    text = repr(value)
+    try:
+        sys.stdout.write(text)
+    except UnicodeEncodeError:
+        bytes = text.encode(sys.stdout.encoding, "backslashreplace")
+        if hasattr(sys.stdout, "buffer"):
+            sys.stdout.buffer.write(bytes)
+        else:
+            text = bytes.decode(sys.stdout.encoding, "strict")
+            sys.stdout.write(text)
+    sys.stdout.write("\n")
+    builtins._ = value  # type: ignore