+++ src/tests/test_console.py
@@ -1,6 +1,7 @@
 import pytest
 from pathlib import Path
 import sys
+import io
 
 sys.path.append(str(Path(__file__).parents[2] / "src" / "pyodide-py"))
 
@@ -23,15 +24,13 @@ def test_stream_redirection():
 
 
 @pytest.fixture
-def safe_stdstreams():
-    stdout = sys.stdout
-    stderr = sys.stderr
+def safe_sys_redirections():
+    redirected = sys.stdout, sys.stderr, sys.displayhook
     yield
-    sys.stdout = stdout
-    sys.stderr = stderr
+    sys.stdout, sys.stderr, sys.displayhook = redirected
 
 
-def test_interactive_console_streams(safe_stdstreams):
+def test_interactive_console_streams(safe_sys_redirections):
 
     my_stdout = ""
     my_stderr = ""
@@ -66,6 +65,12 @@ def test_interactive_console_streams(safe_stdstreams):
     shell.push("print('foobar')")
     assert my_stdout == "foo\nfoobar\n"
 
+    shell.push("print('foobar')")
+    assert my_stdout == "foo\nfoobar\nfoobar\n"
+
+    shell.push("1+1")
+    assert my_stdout == "foo\nfoobar\nfoobar\n2\n"
+
     shell.restore_stdstreams()
 
     my_stdout = ""
@@ -94,3 +99,91 @@ def test_interactive_console_streams(safe_stdstreams):
 
     print("bar")
     assert my_stdout == "foobar\n"
+
+    shell.push("print('foobar')")
+    assert my_stdout == "foobar\nfoobar\n"
+
+    shell.push("import sys")
+    shell.push("print('foobar', file=sys.stderr)")
+    assert my_stderr == "foobar\n"
+
+    shell.push("1+1")
+    assert my_stdout == "foobar\nfoobar\n2\n"
+
+
+def test_repr(safe_sys_redirections):
+    sep = "..."
+    for string in ("x" * 10 ** 5, "x" * (10 ** 5 + 1)):
+        for limit in (9, 10, 100, 101):
+            assert len(
+                console.repr_shorten(string, limit=limit, separator=sep)
+            ) == 2 * (limit // 2) + len(sep)
+
+    sys.stdout = io.StringIO()
+    console.displayhook(
+        [0] * 100, lambda v: console.repr_shorten(v, 100, separator=sep)
+    )
+    assert len(sys.stdout.getvalue()) == 100 + len(sep) + 1  # for \n
+
+
+@pytest.fixture
+def safe_selenium_sys_redirections(selenium):
+    selenium.run("_redirected = sys.stdout, sys.stderr, sys.displayhook")
+    yield
+    selenium.run("sys.stdout, sys.stderr, sys.displayhook = _redirected")
+
+
+def test_interactive_console(selenium, safe_selenium_sys_redirections):
+    def ensure_run_completed():
+        selenium.driver.execute_async_script(
+            """
+        const done = arguments[arguments.length - 1];
+        pyodide.globals.shell.run_complete.then(done);
+        """
+        )
+
+    selenium.run(
+        """
+    from pyodide.console import InteractiveConsole
+
+    result = None
+
+    def displayhook(value):
+        global result
+        result = value
+
+    shell = InteractiveConsole()
+    sys.displayhook = displayhook"""
+    )
+
+    selenium.run("shell.push('x = 5')")
+    selenium.run("shell.push('x')")
+    ensure_run_completed()
+    assert selenium.run("result") == 5
+
+    selenium.run("shell.push('x ** 2')")
+    ensure_run_completed()
+    assert selenium.run("result") == 25
+
+    selenium.run("shell.push('def f(x):')")
+    selenium.run("shell.push('    return x*x + 1')")
+    selenium.run("shell.push('')")
+    selenium.run("shell.push('[f(x) for x in range(5)]')")
+    ensure_run_completed()
+    assert selenium.run("result") == [1, 2, 5, 10, 17]
+
+    selenium.run("shell.push('def factorial(n):')")
+    selenium.run("shell.push('    if n < 2:')")
+    selenium.run("shell.push('        return 1')")
+    selenium.run("shell.push('    else:')")
+    selenium.run("shell.push('        return n * factorial(n - 1)')")
+    selenium.run("shell.push('')")
+    selenium.run("shell.push('factorial(10)')")
+    ensure_run_completed()
+    assert selenium.run("result") == 3628800
+
+    # with package load
+    selenium.run("shell.push('import pytz')")
+    selenium.run("shell.push('pytz.utc.zone')")
+    ensure_run_completed()
+    assert selenium.run("result") == "UTC"